dried up code so hooks and components are now reusable
