+++ containers/ActivitiesList.jsx
@@ -1,26 +1,28 @@
 import React, { useContext } from "react";
 import { Text, View } from "react-native";
-import useUpcomingActivities from "../hooks/useUpcomingActivities";
+import useActivitiesList from "../hooks/useActivitiesList";
 import UserContext from "../context/User";
 import { text } from "../theme";
 
-export default function UpcomingActivitiesList() {
+export default function ActivitiesList({ title, props }) {
   const {
     user: { name },
   } = useContext(UserContext);
 
-  const upcomingList = useUpcomingActivities(name);
-  if (upcomingList !== null && upcomingList.length > 0) {
+  // const organised = useOrganisedActivities(name);
+  const organised = useActivitiesList(props);
+  if (organised !== null && organised.length > 0) {
     return (
       <>
-        <Text style={text.subtitle}>Upcoming Activities</Text>
-        {upcomingList.map((activity) => (
+        <Text style={text.subtitle}>{title}</Text>
+        {organised.map((activity) => (
           <View
             key={activity.id}
             style={{ backgroundColor: "pink", margin: 10, padding: 10 }}
           >
             <Text style={text.body}>{activity.Activity}</Text>
             <Text style={text.body}>{activity.Image}</Text>
+            {/* <Text style={text.body}>{activity.Date}</Text> */}
           </View>
         ))}
       </>
